Ability Fix
This was a pain to get the abilities to show in-game.  For some reason, the Hard Mode ability is the entire brick of text, which is pretty unusual.

Also, both texts end with "\n", so I wasted a lot of time stumbling around until I guessed that particular invisible character.