Update 70001.txt
Explicitly stated the term Life Points before introducing an abbreviation (LP), as was done in the original Japanese text.

Added a sentence explaining how to reduce LP in order to accommodate the contrasting "However" at the start of the next sentence, which explains how to prevent the loss of LP.

Added a comma, since "as long as" marks the beginning of the sentence as a dependent clause. Clauses are separated by a comma when the dependent one comes first.

Removed the last sentence because it is redundant (it restates the previous sentence) and ambiguous (Which player has one, and what does that player have?).