Update 70001.txt
I believe this line of tutorial text is meant to emphasize the fact that the player can attack, not the fact that the attacks can be direct. It was probably written under the assumption the player would already be familiar with Yugioh, where monsters do all the attacking, which means this line is drawing attention to an important difference in game mechanics.

I added a mention of monsters to the last sentence to make it clearer who the point of reference is for the comparison.