Tutorial punctuation
I removed the period because tildes are used consistently as sentence-ending punctuation already, making it redundant. (Note: lots of sentences end in hearts, which are accompanied by separate punctuation marks. Consequently, I do not consider hearts sentence-ending punctuation marks.)

I added commas to separate the clause out of the middle of the sentence. Em-dashes would probably be fine here as well, and I have noticed them used occasionally in this project, but I'm playing it safe with commas here.