Update 10016.txt
I believe this was a mistranslation. Google translate gave me a similar result to the original "despite" version, but it looks like it just inferred the relationship between the two halves of the sentence incorrectly. If anyone better at Japanese wants to review my interpretation, they are welcome to do so.