Eruru description
In addition to the typo, I feel the description could be improved. The existing version somewhat gives the impression that she is intentionally swaying her breasts AND doing other things to draw attention to them, which is probably not the case. I also like adding "unsuspecting" here to emphasize how effectively she lowers the guard of her marks.