Guild Exam Duel and Event First Pass
I made an effort to fix any obvious typos on the first run through this section of the game. I checked for commas that were not escaped and changed them to semicolons. I also sought out terms in all the related files and that require capitalization and made changes where needed. Some instances of Magic- or Permanent Card were changed to use (Continuous) Spell Card instead.

Changing Great Sword Guardian to Greatsword Guardian was done mostly to suit my preference, though there was already a reference to "Greatsword Warrior" on line 86 of 70007.txt.

I made slight changes to the descriptions of the examiner and Greatsword Guardian in an effort to make them more accurate to the original Japanese.

I also examiner's line 21, since "swordswoman" was gendered in a deliberate way and I though "huntress" would match better.