PR Automated Comma Check Feedback Changes
The check indicated that there were issues with this file on lines 23, 159, and 160. The file is less than 160 lines long, so that information is not enough to convey the real nature of the problem to me. I removed trailing commas and whitespace in the file to hopefully resolve the failed check.